requirements.txt
django
pillow
gunicorn
whitenoise
django-import-export